Removed useless ImageUploadData class and moved instances to UploadAction
